% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_missing_data.R
\name{add_missing_data}
\alias{add_missing_data}
\title{Function to add missing data to the LFS}
\usage{
add_missing_data(data)
}
\arguments{
\item{data}{Data frame containing ABS LFS data}
}
\value{
A data frame with additional rows corresponding to the added data
}
\description{
For example, employment is reported at state level for totals and
for full-time, but not for part-time. We calculate part-time.
}
