% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_missing_data.R
\name{add_part_time}
\alias{add_part_time}
\title{Function to add missing part-time employment numbers at state level
Calculated based on totals and full time numbers}
\usage{
add_part_time(data)
}
\description{
Function to add missing part-time employment numbers at state level
Calculated based on totals and full time numbers
}
