% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_and_save_df.R
\name{compress_and_save_df}
\alias{compress_and_save_df}
\title{Compress and save a data frame}
\usage{
compress_and_save_df(df, qs_file)
}
\arguments{
\item{df}{A data frame that is presumed to contain a \code{table_no} column and
other columns.}

\item{qs_file}{Path, incl. filename (with "\code{.qs}" extension) where data
frame is saved.}
}
\value{
\code{TRUE} invisibly on success
}
\description{
Converts any character columns in the data frame to factor, then nests the
data frame (so there's one row per \code{table_no}), then saves
using \code{qs::qsave()}.
}
\details{
Any character columns (other than \code{table_no}) will be saved as
factors. Loading the data frame with \code{load_data()} converts factor columns back
to character.
}
\examples{
\dontrun{

df <- readabs::read_abs("6345.0")
compress_and_save_df(df, "wpi.qs")

# Then load the data
loaded_df <- load_data("wpi.qs")
}

}
