% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{djpr_connect}
\alias{djpr_connect}
\title{DJPR PostgreSQL Database Connections}
\usage{
djpr_connect(
  db = c("opendata", "official_sensitive"),
  user = c("open", "super"),
  use_config = FALSE,
  ...
)
}
\arguments{
\item{...}{additional arguments passed to \code{\link[config:get]{config::get()}}}
}
\description{
Connect to the DJPR AWS PostgreSQL server and databases. All
users must have authentication details stored as environment variables.
You can add these via \code{\link[usethis:edit]{usethis::edit_r_environ()}}
}
\examples{
\dontrun{
djpr_connect()
}
}
