% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_and_unzip.R
\name{dl_and_read}
\alias{dl_and_read}
\title{Download files with dl_and_unzip() and read using read_abs_local()
Downloads a zip to a temp file, unzips the contents to \code{dest_dir},
loads the contents using \code{readabs::read_abs_local_dir()}}
\usage{
dl_and_read(
  catalogue_string,
  dest_dir = file.path("data-raw", "raw-data", catalogue_string)
)
}
\arguments{
\item{catalogue_string}{eg. "labour-force-australia"}

\item{dest_dir}{Directory in which}
}
\description{
Download files with dl_and_unzip() and read using read_abs_local()
Downloads a zip to a temp file, unzips the contents to \code{dest_dir},
loads the contents using \code{readabs::read_abs_local_dir()}
}
