% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfs_pivot.R
\name{get_lfs_pivot}
\alias{get_lfs_pivot}
\title{Download and tidy ABS pivot cubes}
\usage{
get_lfs_pivot(
  cube,
  catalogue_string = "labour-force-australia-detailed",
  path = Sys.getenv("R_READABS_PATH", unset = tempdir()),
  sheet = "Data 1",
  col_names = TRUE
)
}
\arguments{
\item{cube}{Unique fragment of datacube filename, such as "EQ03"}

\item{catalogue_string}{Such as "labour-force-australia-detailed". See
\code{?readabs::show_available_catalogues} and \code{?readabs::search_catalogues}.}

\item{path}{Location to store the raw data cube}

\item{sheet}{Name of the Excel sheet containing the data for the pivot cube}

\item{col_names}{Passed to \code{col_names} argument of \code{readxl::read_excel()}.
If \code{TRUE}, the default, the text in the spreadsheet will be used as the
column names; a character vector can be supplied instead.}
}
\description{
Download and tidy ABS pivot cubes
}
\examples{
\dontrun{
get_lfs_pivot("EQ03")
}
}
