% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lfs_lookup}
\alias{lfs_lookup}
\title{Information about series in the ABS Labour Force Survey}
\format{
A data frame with >20,000 rows and 27 variables
}
\usage{
lfs_lookup
}
\description{
For each time series ID in the labour force survey (incl. detailed monthly
and quarterly releases), includes information about
}
\keyword{datasets}
