% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load a nested data frame saved as a \code{qs} file}
\usage{
load_data(qs_file)
}
\arguments{
\item{qs_file}{The path, incl. filename and extension, to a \code{qs} file.
File is presumed to contain a dataframe with a nested column called \code{data}.
No checking is performed.}
}
\value{
A nested tibble
}
\description{
Load a nested data frame saved as a \code{qs} file
}
\details{
Any columns in the nested \code{data} column that are factors will be
converted to character.

Intended for use with dataframes saved using \code{compress_and_save_df()}.
}
\examples{
\dontrun{
load_data("qs_file.qs")
}
}
