% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abs_if_updated.R
\name{read_abs_if_updated}
\alias{read_abs_if_updated}
\title{Download catalogue number worth of ABS time series data if it has been
updated since the version stored locally}
\usage{
read_abs_if_updated(
  cat_no = NULL,
  path = here::here("data-raw", "abs-ts"),
  include_trend = FALSE,
  ...
)
}
\arguments{
\item{cat_no}{character, such as "6202.0" or "6345.0"}

\item{path}{default is \code{file.path("data-raw", "abs-ts")}}

\item{include_trend}{logical; default is \code{FALSE}. Should 'trend' data be
included in the data frame?}

\item{...}{arguments passed to \code{readabs::read_abs()}}
}
\description{
Checks the latest observation date in a local ABS time series file; compares
that to the latest date for the catalogue number on the ABS website. If the
remote version is newer than the local version, download the catalogue
number using \code{readabs::read_abs()}
}
\examples{
\dontrun{
read_abs_if_updated("6345.0")
}
}
