% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abs_if_updated.R
\name{read_abs_local_dir}
\alias{read_abs_local_dir}
\title{Read all time series spreadsheets from a given directory}
\usage{
read_abs_local_dir(dir)
}
\arguments{
\item{dir}{Directory containing ABS time series spreadsheets}
}
\value{
A tbl_df
}
\description{
Read all time series spreadsheets from a given directory
}
\details{
uses \code{readabs::read_abs_local()}
}
