% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_jobactive.R
\name{read_jobactive}
\alias{read_jobactive}
\title{Download and tidy jobactive caseload data from the LMIP}
\usage{
read_jobactive(file = tempfile(fileext = ".xlsx"))
}
\arguments{
\item{file}{Path, including filename and extension, to the location where
the jobactive Excel file should be downloaded.}
}
\value{
A \code{tbl_df}
}
\description{
The Commonwealth Government publishes jobactive caseload data publicly on the
Labour Market Information Portal. This function downloads and tidies that
data.
}
\examples{
\dontrun{
read_jobactive()
}
}
