% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_salm.R
\name{read_salm}
\alias{read_salm}
\title{Download and tidy SALM data from the LMIP}
\usage{
read_salm(file = tempfile(fileext = ".xlsx"))
}
\arguments{
\item{file}{Path, including filename and extension, to the location where
the small area labour markt Excel file should be downloaded.}
}
\value{
A \code{tbl_df}
}
\description{
The Commonwealth Government publishes small area labour market data publicly on the
Labour Market Information Portal. This function downloads and tidies that
data.
}
\examples{
read_salm()
}
