% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_summarise.R
\name{ts_summarise}
\alias{ts_summarise}
\title{Summarise a data frame containing ABS time series}
\usage{
ts_summarise(df)
}
\arguments{
\item{df}{Data frame containing time series}
}
\value{
Data frame with the following columns
\itemize{
\item {latest_val}{Latest value}
\item {latest_date}{Date of latest value}
\item {latest_month}{Month of latest value}
\item {latest_year}{Year of latest value}
\item {d_latest}{Change between previous and latest period}
\item {d_year}{Change between latest observation and prior year}
}
}
\description{
Summarise a data frame containing ABS time series
}
