% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_lfs_data.R
\name{filter_lfs_data}
\alias{filter_lfs_data}
\title{Filter a dataframe based on series ID}
\usage{
filter_lfs_data(
  indicator,
  series_type = "seasonally adjusted",
  ...,
  df = .env$dash_data
)
}
\arguments{
\item{indicator}{Indicator name, such as 'unemployment rate'.
See \code{?find_lfs_series()} for possible values.}

\item{series_type}{Default is 'seasonally adjusted'.}

\item{...}{arguments passed to \code{find_lfs_series()} and used to
filter \code{df}.}

\item{df}{Data frame, presumed to be nested, and to
contain a \code{series_id} column with ABS Time Series IDs and
a \code{data} column that is nested. Default is \code{dash_data}, the
data frame used by \code{{djprlabourdash}}.}
}
\value{
a data frame (tibble) filtered based on the filtering
conditions specified in \code{...}
}
\description{
Series ID found using \code{find_lfs_series()}
}
\examples{
\dontrun{
abs_6202 <- readabs::read_abs("6202.0", 5)
filter_lfs_data("unemployment rate", state = "victoria", df = abs_6202)
}
}
