% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfs_pivot.R
\name{get_all_lfs_pivots}
\alias{get_all_lfs_pivots}
\title{Get all pivot tables for which functions have been created
Need to manually add them}
\usage{
get_all_lfs_pivots(
  path = Sys.getenv("R_READABS_PATH", unset = tempdir()),
  all_states = FALSE
)
}
\arguments{
\item{path}{Path to directory in which to store downloaded file(s)}

\item{all_states}{logical, \code{FALSE} by default. When \code{FALSE}, states other than
Victoria will be removed.}
}
\description{
Get all pivot tables for which functions have been created
Need to manually add them
}
