% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load a data frame saved as a \code{qs} file and convert factors to strings}
\usage{
load_data(qs_file)
}
\arguments{
\item{qs_file}{The path, incl. filename and extension, to a \code{qs} file.}
}
\value{
A tibble
}
\description{
Load a data frame saved as a \code{qs} file and convert factors to strings
}
\details{
Any columns that are factors will be
converted to character.

Intended for use with dataframes saved using \code{compress_and_save_df()}.
}
\examples{
\dontrun{
load_data("qs_file.qs")
}

}
